Added plotting to show wins vs losses in xy space
