+++ README.md
@@ -1,2 +1,6 @@
 # nhl-predictor
 Repository for testing math models to predict outcomes of games based on numerous predictor variables. First project using JetBrains new DataSpell IDE
+
+## TODO
++ Accessing csv is everywhere. Split into repository pattern and prepare for database access
+instead of data access in business logic
